% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{asim}
\alias{asim}
\title{Annular simulations}
\format{
A list with the following components:
\itemize{
  \item{mar:}{ APSF simulation for the maritime aerosol type.}
  \item{con:}{ APSF simulation for the continental aerosol type.}
  \item{urb:}{ APSF simulation for the urban aerosol type.}
  \item{ray:}{ A list with APSFs of Rayleigh scattering for different surface pressures.}
  \itemize{
    \item{p1100:}{ Rayleigh APSF simulation at 1100 mbar.}
    \item{p1013:}{ Rayleigh APSF simulation at 1013.25 mbar.}
    \item{p0750:}{ Rayleigh APSF simulation at 750 mbar.}
    \item{p0500:}{ Rayleigh APSF simulation at 500 mbar.}
  }
}
}
\source{
Generated with the apsfs library.
}
\usage{
asim
}
\description{
Contains the atmospheric point spread functions (APSFs) for nadir vieweing 
sensors, simulated for the different standard aerosol models of the 6SV 
radiative transfer code: maritime (mar), continental (con) and urban (urb). 
Simulations were carried out without Rayleigh contribution, and with an 
aerosol scale height of 2 km. The APSFs for Rayleigh alone (ray) are also 
available at four pressure levels: 1100 mbar (p1100), 1013.25 mbar (p1013), 
750 mbar (p0750) and 500 mbar (p0500). Sensor was positioned at TOA, with 
nadir view angle and infinitesimal field of view. The vertical optical 
thickness for all aerosol simulations was 0.5, while for Rayleigh, was the
Rayleigh optical thickness at the given surface atmospheric pressure at 450 
nm. It is noted, however, that dependence os the normalized APSF on optical 
thickness is small. By construction, aerosol APSF have no pressure 
dependence. All simulations are for a sensor altitude of 800 km, but altitude
dependence is only significant at altitudes lower than 10 km (aircraft, 
drones).
}
\details{
Simulations run with the annular accumulator geometry.
}
\keyword{datasets}
