% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scat.R
\name{calc_cdf}
\alias{calc_cdf}
\title{Aerosol CDF Look Up Table}
\usage{
calc_cdf(ph, xout, type = "fixed")
}
\arguments{
\item{ph}{Phase function table at single wavelength.}

\item{xout}{Angle points for calculation.}

\item{type}{One of "fixed", "dynamic". See Details.}
}
\description{
Calculates a Look Up Table (LUT) for the aersol cumulative distribution 
function to be used with the Monte Carlo calculations.
}
\details{
This function calculates the cumulative density function of aerosol 
phase functions. It can use a fixed integration grid based on xout (must be 
very fine resolution!) or an adaptive integration (any xout). Note that the 
convention in atmospheric optics is to have the PF not normalized by 4pi 
steradian, so it is unitless instead of 1/sr as in ocean optics. The integral 
to 1 is:
integral_0^2pi integral_0^pi (PF(theta) / 4pi) * sin(theta) * dtheta * dphi = 1
such that:
integral_0^pi PF(theta) / 2 * sin(theta) * dtheta = 1

Due to small numerical errors, the phase functions are renormalized to 1 
before returning.
}
\examples{
psi_out <- c(0, 10^seq(log10(0.00001), log10(180), length.out = 1000))
calc_cdf(continental_ph_6sv[, c(1, 8)], psi_out)

}
