% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scat.R
\name{cdf_lut}
\alias{cdf_lut}
\title{Aerosol CDF Look Up Table}
\usage{
cdf_lut(ph, xout, method = "fixed")
}
\arguments{
\item{ph}{Phase function table at single wavelength.}

\item{xout}{Angle points for calculation.  See Details.}

\item{method}{One of "fixed", "adaptative". See Details.}
}
\value{
A numeric vector with the aerosol scattering CDF.
}
\description{
Calculates a Look Up Table (LUT) for the aersol cumulative distribution 
function to be used with the Monte Carlo calculations.
}
\details{
This function calculates the cumulative distribution function of 
aerosol phase functions between 0 and pi. It can use a fixed integration grid 
based on xout (must be fine angular grid resolution!) or an adaptive 
integration (any xout) based on \code{integrate}. Regardless of the method 
used, a fine resolution is recommended to reduce bias on the PSF simulation. 
If xout is not specified, the defaault is to use 1000 log10 spaced values 
between 0 and pi. If xout is provided, it must include 0 and pi.

For the integration, a linear interpolation (\code{approx}) is used in log10
space for the aerosol phase function. Method 'fixed' uses fixed integration 
steps with quadrature rule.

Note that the convention in atmospheric optics is to have the PF not 
normalized by 4pi steradian, so it is unitless instead of 1/sr as in ocean 
optics. The integral to 1 is:

integral_0^2pi integral_0^pi (PF(theta, phi) / 4pi) * sin(theta) * dtheta * dphi = 1,

which in isotropic conditions (lack of azimuthal dependence) simplified to:

integral_0^pi (PF(theta) / 2) * sin(theta) * dtheta = 1

Due to small numerical errors, the phase functions are renormalized to 1 
before returning.
}
\examples{
cdf_lut(continental_ph_6sv[, c(1, 8)])

}
