% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{continental_ph_6sv}
\alias{continental_ph_6sv}
\alias{maritime_ph_6sv}
\alias{urban_ph_6sv}
\title{6SV Standard Aerosol Phase Functions}
\format{
A data frame with 83 rows and 21 variables, the first being the 
angular reference (degrees) and the phase function in 20 wavelengths.

An object of class \code{data.frame} with 83 rows and 21 columns.

An object of class \code{data.frame} with 83 rows and 21 columns.
}
\source{
6SV code \url{http://6s.ltdri.org/}
}
\usage{
continental_ph_6sv

maritime_ph_6sv

urban_ph_6sv
}
\description{
A dataset containing the scattering phase functions of three standard 
aerosol models included in 6SV (maritime, continental and urban).
}
\keyword{datasets}
