% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cos2sph.R
\name{cos2sph}
\alias{cos2sph}
\title{Convert form directional cosines to spherical directions}
\usage{
cos2sph(cdir)
}
\arguments{
\item{cdir}{A vector containing the directional cosines (dx/dl, dy/dl, dz/dl).}
}
\value{
A vector of length two containing the polar angle (psi) and the 
azimuthal angle (phi).
}
\description{
Calculates the polar spherical directions from directional cosines.
}
