% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psf_m.R
\name{cum_psf}
\alias{cum_psf}
\title{Cumulative PSF}
\usage{
cum_psf(psfi, norm = TRUE)
}
\arguments{
\item{psfi}{An object produced by the function \code{mc_psf}.}

\item{norm}{Logical. Should the PSF be normalize to integrate to unity?}
}
\value{
A numeric vector if the input simulation is in annular geometry and 
a matrix if is in sectorial or grid geometry.
}
\description{
Calculates the uni (annular) or bidimensional (sectorial, grid) cumulative 
PSF.
}
\details{
The Monte Carlo simulation returns the area integral PSF on the 
intervals defined by the break points of the accumulator. This function sum 
those contributions evaluating the cumulative distribution at the break 
points.
}
