% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivative.R
\name{.get_d2psf_dxdy}
\alias{.get_d2psf_dxdy}
\title{Calculates the mixed partial derivative:  d^2 * z / dx * dy for a scalar 
valued function (z) of two arguments (x, y). This is the bidimensional 
density function, the underlying process integrated by Monte Carlo, 
sub-sequentially used to calculate the bidimensional cumulative 
distribution function fitted by fit_sectorial or fit_grid.}
\usage{
.get_d2psf_dxdy(func, x1, x2, ...)
}
\description{
The code is modified from numDeriv:::genD. The original function is generic 
and vectorized in the sense of a vector valued function. It was adapted here 
to solve a scalar valued function at several coordinates at the same time for 
reasons of computational efficiency. The modifications below make this 
function specific to the problem at hand, in particular the way the 
predict_sectorial subfunctions are expected to work.
}
\details{
It is based on the numDeriv R package source code version 2016.8-1.1, that is 
licensed under GPL-2. This package is also licensed as GPL-2 and a copy of the
license is provided in the root directory of the package. Modifications made 
on 2020-03-14.
}
