% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc.R
\name{.movpht_m}
\alias{.movpht_m}
\title{Ancillary function to move photon in stratified atmosphere:}
\usage{
.movpht_m(pht, s, atm, cid)
}
\description{
Ancillary function to move photon in stratified atmosphere:
}
