% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psf_m.R
\name{dpsf}
\alias{dpsf}
\title{PSF normalized per bin area}
\usage{
dpsf(psfm, norm = TRUE)
}
\arguments{
\item{psfm}{An object produced by the function \code{mc_psf}.}

\item{norm}{Logical. Should the PSF be normalized to sum to unity?}
}
\value{
A vector with the average per area PSF in each bin.
}
\description{
Normalized the bin integral per area of bin used in the Monte Carlo 
accumulation.
}
\details{
PSF are normalized per area of each bin depending on geometry. If 
norm is TRUE, PSF is normalize to sum to 1 before calculation and the 
integral of the area normalized PSF will be 1.

Note that the bin containing data from the last requested distance to 
infinite will have a value of zero per area PSF.
}
