% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psf_m.R
\name{fit_annular_psf}
\alias{fit_annular_psf}
\title{PSF annular cumulative integral model}
\usage{
fit_annular_psf(psfm, press = NULL, norm = TRUE, nstart = 10)
}
\arguments{
\item{psfm}{A list of objects created by the function \code{mc_psf}.}

\item{press}{Pressure in mbar or NULL. See details.}

\item{norm}{Logical. Should the PSF be normalized to sum to unity?}

\item{nstart}{Number of random start positions for the multidimension 
function optimization.}
}
\value{
A list with the following components: 
"type":         type of the fitted model; 
"coefficients": named vector with the 6 coefficients;
"mare":         the minimum mean absolute relative error of the model fit; 
"mare_seq":     a sorted sequence of MARE of different random starts;
"convergence":  convergence code from \code{optim};
"press_dep":    logical flag indicating if there was pressure dependency; 
"press_rng":    the range of pressure values if fitted with pressure 
                dependency.
}
\description{
Fits an exponential model to the radius cumulative integral of a PSF in 
annular geometry of a given scatter model, possibly with pressure dependency.
}
\details{
The function calculates the (normalized) cumulative integral of the annular 
PSF and fits the following model by optimization:

F(r) = c1 - (c2 * e^(c3 * r) + c4 * e^(c5 * r) + c6 * e^(c7 * r)),

c1 = x1,
c2 = x2 / p,
c3 = x3 / p,
c4 = (c1 - c2) * x4,
c5 = x5,
c6 = (c1 - c2) * (1 - x4),
c7 = x6,
p  = press / 1013.25,
and x1 to x5 are model parameters. Note that x1 = total diffuse transmittance
(or 1, if norm = TRUE) and that c2 + c4 + c6 = c1.

If there is no pressure dependency (e.g., aerosol only), 'press' does not 
need to be specified. Internally, it will be set to 1013.25 mbar, p will 
equal unity and the pressure flag will be set to FALSE in the fit object 
metadata. The prediction function \code{predict_annular}, will track 
dependency on pressure and model domain.

Note that adding multiple models in the input list will only be meaningfull 
if they are simulations for the same scatter at different surface pressure 
levels. If psfm is a list with more than one component and press is 
specified, norm is automatically set to TRUE.

The cumulative annular integral PSF is as known as "environmental function".
The optimization is made with mean absolute relative error (MARE) as the 
error function.

Note that 'PSF' in this package refers to the diffusely transmitted photons 
only. Photons from direct transmission are simulated and are part of the PSF,
but not included in the fit calculations.
}
\examples{

psfm <- list(
 ray_res003_v00_p1100_annular,
 ray_res003_v00_p1013_annular,
 ray_res003_v00_p0750_annular,
 ray_res003_v00_p0500_annular
)
press <- rep(c(1100, 1013.25, 750, 500), each = length(ray_res003_v00_p0500_annular$bin_mid))
opt   <- fit_annular_psf(psfm, press = press, norm = TRUE, nstart = 30)
opt
plot(opt$mare_seq, xlab = "Iteration", ylab = "MARE")

}
