% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psf_m.R
\name{fit_radial_psf}
\alias{fit_radial_psf}
\title{PSF annular cumulative integral model}
\usage{
fit_radial_psf(psfm, norm = TRUE)
}
\arguments{
\item{psfm}{An object produced by the function \code{mc_psf}.}

\item{norm}{Logical. Should the PSF be normalized to sum to unity?}
}
\value{
A list with the follwoing components: "type": type of the fitted model, 
"coefficients": named vector with the 5 coefficients, and "mare" with the 
value of the mean absolute relative error of the model fit.
}
\description{
Fits an exponential model to the annular cumulative integral of a radial grid
PSF.
}
\details{
The code calculates the cummulative integral of the radial PSF and fits the 
following model by optimization:

F(r) = c1 - (c2 * e^(c3 * r) + c4 * e^(c5 * r))

The cummulative annular integral PSF is as known as "environmental function".
The optimization is made with mean absolute relative error (MARE) as the 
error function.

Note that 'PSF' in this package refers to the diffusely transmitted photons 
only. Photons from direct transmission are not included in the caluclations.
}
