% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atm_prfl.R
\name{get_mol_par}
\alias{get_mol_par}
\title{Get molecular absorption profile}
\usage{
get_mol_par(atm, lambda)
}
\arguments{
\item{atm}{Profile of atmospheric temperature, pressure, water vapor and 
ozone.}

\item{lambda}{Wavelength (nm) of observation.}
}
\value{
A vector with the molecular absorption profile at the same heights as
the atmospheric profile.
}
\description{
Calculates the molecular absorption profile for a given atmosphere.
}
\details{
Not implemented. At the moment will return a vector of zeros (zero
absorption) with equal length to the atmospheric profile.
}
