% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atm_prfl.R
\name{get_opt_atm_prfl}
\alias{get_opt_atm_prfl}
\title{Get optical properties of atmospheric layers}
\usage{
get_opt_atm_prfl(
  atm,
  tau_aer,
  H_aer,
  w0_aer,
  tau_ray_z,
  a_mol_z,
  nlayers = 1000
)
}
\arguments{
\item{atm}{A data.frame with columns "Z" (Km height from surface) and 
"Pressure" (mbar).}

\item{tau_aer}{Aerosol optical thickness (unitless), [0,Inf).}

\item{H_aer}{Aerosol scale height (Km), [0,Inf).}

\item{w0_aer}{Aerosol single scattering albedo (unitless), [0,1].}

\item{tau_ray_z}{Rayleigh optical thickness (unitless), [0,Inf).}

\item{a_mol_z}{Molecular absorption profile at the same Z and Pressure 
points of atm.}

\item{nlayers}{Number of layers of the output.}
}
\value{
A data.frame with the average optical properties per layer. Note that 
the first two columns are the the break points of the atmosphere layers in 
tau and Km. So the last row will always be NA for all average properties,
since the last atmosphere layer is nrow - 1.
}
\description{
Calculates the average optical properties per atmosphere layer for a given 
atmospheric profile.
}
\details{
Linear interpolation is used for the atmospheric profile of pressure and 
molecular absorption. The atmosphere is returned as nlayers between the 
surface (0 km) and the top of the input atmospheric profile.

Pseudo-code:
1 - Get break points of atmosphere layers;
2 - Based on the exponential profile, integrate the aerosol extinction from 
    TOA to surface.
3 - Calculate average aerosol properties per layer.
4 - Based on the pressure profile, get the Rayleigh optical depth profile.
5 - Calculate average Rayleigh scattering per layer.
6 - Integrate molecular absorption from TOA to surface.
7 - Calculate the average molecular absorption per layer.
8 - Calculate the average extinction per layer and single scattering albedo.
9 - Sum the total optical depth from TOA per layer.
}
