% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gsim}
\alias{gsim}
\title{Grid simulations}
\format{
A list with the following components:
\itemize{
  \item{conv00:}{ APSF simulation for the continental aerosol type at 0 degree view angle;}
  \item{conv30:}{ APSF simulation for the continental aerosol type at 30 degree view angle;}
  \item{conv60:}{ APSF simulation for the continental aerosol type at 60 degree view angle.}
}
}
\source{
Generated with the apsfs library.
}
\usage{
gsim
}
\description{
Contains the atmospheric point spread functions (APSFs) for nadir, 30 and 60 
degrees vieweing sensors, simulated for the continental (con) standard aerosol 
model of the 6SV radiative transfer code. Simulations were carried out 
without Rayleigh contribution, and with an aerosol scale height of 2 km.
Sensor was positioned at TOA, with an infinitesimal field of view. The 
vertical optical thickness for all aerosol simulations was 0.5. It is noted, 
however, that dependence os the normalized APSF on optical thickness is small. 
By construction, aerosol APSF have no pressure dependence. All simulations 
are for a sensor altitude of 800 km, but altitude dependence is only 
significant at altitudes lower than 10 km (aircraft, 
drones).
}
\details{
Simulations run with the grid accumulator geometry.
}
\keyword{datasets}
