% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scat.R
\name{n_air}
\alias{n_air}
\title{Dispersion formula for the refractive index of air}
\usage{
n_air(lambda)
}
\arguments{
\item{lambda}{Wavelength in vacuum (nm).}
}
\value{
A numeric vector with the real part of refractive index of standard 
dry air (unitless).
}
\description{
This function calculates the real part of the refractive index of standard 
dry air relative to vaccum. See details.
}
\details{
The dispersion model implemented is that by Ciddor (1996), for the 
refractive index of standard dry air. It is reference for dry air (0 
moisture), 15 ºC, 101.325 Pa and 450 ppm CO2. This equation is recomended by 
Zhang et al. (2009) for the convertion of (saline) water refractive index 
relative to air to that relative to vacuum.

Ciddor (1996) also provides equations to account for different levels of CO2 
and water vapor. Those are not implemented in the current version.

The valid range is 200 nm to 1100 nm. If the refractive index is requested 
outside this range, the function will return a warning.
}
\examples{
# Get the refractive index for teh visible range: 
n_air(400:700) 

}
\references{
Ciddor, P. E. 1996. Refractive index of air: new equations for the visible 
and near infrared. Applied Optics 35, 9, 1566-73. DOI: 10.1364/AO.35.001566

Zhang, X.; Hu, L.; He, M.-X. 2009. Scattering by pure seawater: effect of 
salinity. Optics Express 17, 7, 5698-710. DOI: 10.1364/OE.17.005698
}
