% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psf_m.R
\name{predict_annular}
\alias{predict_annular}
\title{Predict annular PSF fitted model}
\usage{
predict_annular(r, fit, cum = FALSE)
}
\arguments{
\item{r}{The radius distances (km) at which the model should be evaluated.}

\item{fit}{A model fit from \code{fit_annular_psf}.}

\item{cum}{Logical. Should the cummulative PSF be returned? Defaults to FALSE.}
}
\value{
A numeric vetor with the PSF (or cummulative PSF).
}
\description{
Solves a fitted model from \code{fit_annular_psf} for the PSF of the radial 
cummulative PSF at requested radius.
}
\details{
If cum = FALSE, the derivative of the cummulative PSF model fit will 
be returned, which is the PSF at the desired radius.
}
