% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psf_m.R
\name{predict_grid}
\alias{predict_grid}
\title{Predict PSF grid from fitted model}
\usage{
predict_grid(
  f_aer,
  f_ray,
  tray,
  taer,
  ext = 0,
  res = 0.03,
  press = NULL,
  view = 0
)
}
\arguments{
\item{f_aer}{Grid or annular fit to Rayleigh PSF}

\item{tray}{Transmittance due to Rayleigh}

\item{taer}{Transmittance due to aerosols}

\item{ext}{Maximum spatial extent at geom resolution (km).}

\item{res}{Resolution (km), [0,10].}

\item{press}{Pressure in mbar or NULL.}
}
\description{
Predicts a PSF grid from annular or grid fitted models.
}
