% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rayleigh_cdf}
\alias{rayleigh_cdf}
\title{Rayleigh Cumulative Distribution Function}
\format{A data frame with 1000 rows and 2 variables, the first being the 
angular reference (radians) and the second the CDF.}
\source{
6SV code \url{http://6s.ltdri.org/}
}
\usage{
rayleigh_cdf
}
\description{
A dataset containing the Rayleigh CDF at 0.18 degree resolution.
}
\keyword{datasets}
