% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scat.R
\name{rayleigh_od}
\alias{rayleigh_od}
\title{Calculate Rayleigh optical depth}
\usage{
rayleigh_od(atm, lambda, co2 = 400, lat = 45)
}
\arguments{
\item{atm}{Atmosphere profile, see Details.}

\item{lambda}{Wavelength (nm), [200,1100].}

\item{co2}{CO2 concentration (ppm), [0,1E6].}

\item{lat}{Geodetic latitude (degrees), [-90,90].}
}
\value{
A vector with the same length as the number of rows of atm with the Rayleigh
optical depth.
}
\description{
Calculates the Rayleigh optical depth depending on pressure and height above 
sea level.
}
\details{
The atmospheric profile must be a data frame with at least two 
columns: Z, with height above sea level (km); and Pressure, with atmospheric
pressure (mbar).

Note that all arguments should be length = 1 for atm with multiple rows 
(expected). For atm with a single row (e.g., sea level), one of the arguments
can have length > 1.
}
\examples{
rayleigh_od(us62, lambda = 400, co2 = 400, lat = 45)
rayleigh_od(data.frame(Z = 0, Pressure = 1013.25), lambda = 400:700, 
            co2 = 400, lat = 45)

}
\references{
Bodhaine, B.A.; Wood, N.B.; Dutton, E.G.; Slusser, J.R. 1999. On Rayleigh 
optical depth calculations. Journal of Atmospheric and Oceanic Technology 16, 
11, 1854-1861. DOI: 10.1175/1520-0426(1999)016<1854:ORODC>2.0.CO;2
}
