% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_grid.R
\name{rotate_grid}
\alias{rotate_grid}
\title{Rotate azimuth of grid simulations}
\usage{
rotate_grid(psf, vaz)
}
\arguments{
\item{psf}{A grid psf object.}

\item{vaz}{Viewing azimuth (rad).}
}
\description{
Rotate the grid simulation to the requested viewing azimuth.
}
