% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tau_ray_6sv}
\alias{tau_ray_6sv}
\title{6SV Rayleigh Optical Thickness Coefficients at surface pressure}
\format{A data frame with 20 rows and 2 variables, the first being the 
wavelength reference (nm) and the second the optical thickness.}
\source{
6SV code \url{http://6s.ltdri.org/}
}
\usage{
tau_ray_6sv
}
\description{
A dataset containing the optical thickness of Rayleigh scattering at selected
wavelengths of the 6SV aerosol data grid. The surface pressure is 1013.25 
mbar.
}
\keyword{datasets}
