% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updatecdir.R
\name{update_cdir}
\alias{update_cdir}
\title{Update directional cosines}
\usage{
update_cdir(psi, phi, cdir)
}
\arguments{
\item{psi}{Polar scattering angle (radians), [0,pi].}

\item{phi}{Azimuthal scattering angle (radians), [0,2*pi].}

\item{cdir}{Directional cosines (unitless; dx/dl, dy/dl, dz/dl).}
}
\value{
A vector of length 3 containing the new travelling directional 
cosines (dx/dl, dy/dl, dz/dl) of the photon package.
}
\description{
Rotates the reference frame of scattering to the reference frame of the 
photon package previous direction.
}
