% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us62}
\alias{us62}
\title{US62 Standard Atmosphere}
\format{A data frame with 34 rows and 5 variables:
\describe{
  \item{Z}{Height from surface (km)}
  \item{Pressure}{Pressure (mbar)}
  ...
}}
\source{
6SV code \url{http://6s.ltdri.org/}
}
\usage{
us62
}
\description{
A dataset containing model profiles of temperature, pressure, water vapor and
ozone concentration.
}
\keyword{datasets}
