% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zernike.R
\name{zrnk_exp}
\alias{zrnk_exp}
\title{Calculate the Zernike polynomials}
\usage{
zrnk_exp(ord, rho, phi, thetav, sym = FALSE)
}
\arguments{
\item{ord}{Maximum radial order.}

\item{rho}{Radius, [0, 1].}

\item{phi}{Azimuth (radians), [0, 2*pi] or [-pi, pi]}

\item{thetav}{View angle (radians), [0, pi/2]}

\item{sym}{Logical. For Zernike products, if it should respect Helmholtz 
symetry.}
}
\value{
A matrix with Zernike terms per row and positions per column.
}
\description{
This function calculates the Zernike polynomials for the unit disk, or for 
the product of the two unit disks possibly respecting Helmholtz symetry 
relations.
}
\details{
If thetav is missing, the function will return the Zernike 
terms up with radial order up to ord and all azimuthal orders. If thetav is 
present, the same caluclation will be performed for the product of two 
Zernike polynomials for isotropic conditions, i.e., azimuthal dependency is 
given only by the relative azimuth between pixel center and view vector 
projection on the surface. If sym is true, a more complex model is returned
that respects Helmholtz symetry and is defined as:
N * [Rkm(rhoi) * Rkn(rhos) + Rkm(rhoi) * Rkn(rhor)] * cos(k * Delta_phi)
where N is the normalization factor and R are the radial polynomials of 
azimuthal order k and radial orders m and n.

Zernike calculations implemented in C. Vectorized use is encouraged, but rho 
and phi should have the same length.
}
