% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zernike.R
\name{ozernike}
\alias{ozernike}
\title{Calculates the Radial Zernike Polynomials up to a given radial order}
\usage{
ozernike(rho, phi, thetav, ord)
}
\arguments{
\item{rho}{Radius, [0, 1].}

\item{phi}{Azimuth, [0, 2pi].}

\item{thetav}{View angle, [0, pi/2].}

\item{order}{Maximum radial order.}
}
\description{
Calculates the Radial Zernike Polynomials up to a given radial order
}
