% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psf_m.R
\name{predict_annular}
\alias{predict_annular}
\title{Predict annular PSF fitted model}
\usage{
predict_annular(r, fit, type = c("psf", "dpsf", "cumpsf"), press = NULL)
}
\arguments{
\item{r}{The radius distances (km) at which the model should be evaluated.}

\item{fit}{A model fit from \code{fit_annular_psf}.}

\item{type}{Type of prediction: 'psf', 'dpsf', or 'cumpsf'. See Details.}

\item{press}{Pressure in mbar or NULL. See details.}
}
\value{
A numeric vetor with the PSF, dPSF/dArea or cumulative PSF.
}
\description{
Solves a fitted model from \code{fit_annular_psf} for the PSF of the radial 
cummulative PSF at requested radius.
}
\details{
If type = cumpsf, the model fit to the cumulative PSF will be 
evaluated at the desired radius points. If type = dpsf, the area derivative 
the PSF dPSF/dArea is returned. If type == psf, quadrature is used to on the 
average area derivative of the annulus and scaled by the area of the annulus. 
Note that in this case, r will be sorted and the returned values are for the 
mid points of the input vector of radius, so will have a length of 
length(r) - 1. Default is to return the PSF.
}
