% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_grid.R
\name{predict_grid}
\alias{predict_grid}
\title{Predict SAF or PSF grid from fitted model}
\usage{
predict_grid(
  f_aer,
  f_ray,
  wray,
  waer,
  ext = 0,
  res = 0.03,
  press = NULL,
  tpred = NULL,
  vaz = pi/2
)
}
\arguments{
\item{f_aer}{Grid or annular fit to Rayleigh SAF or PSF}

\item{wray}{Weight (transmittance or reflectance) due to Rayleigh}

\item{waer}{Weight (transmittance or reflectance) due to aerosols}

\item{ext}{Maximum spatial extent at geom resolution (km).}

\item{res}{Resolution (km), [0,10].}

\item{press}{Pressure in mbar or NULL. For annular fits only.}

\item{tpred}{Third predictor level or NULL. For sectorial fits only.}

\item{vaz}{Viewing azimuth (rad). For sectorial geometry only.}
}
\description{
Predicts a SAF or PSF grid from annular or grid fitted models.
}
\details{
The grid for each scatter type are calculated individually, then their 
weighted average is calculated based on the weights for Rayleigh and aerosol.
The weights are given by the spherical albedo (for SAF) and diffuse upward 
transmittance (for PSF).
}
