% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_sectorial.R
\name{predict_sectorial}
\alias{predict_sectorial}
\title{Predict sectorial PSF fitted model}
\usage{
predict_sectorial(
  r,
  a = NULL,
  fit,
  type = c("psf", "dpsf", "cpsf"),
  tpred = NULL
)
}
\arguments{
\item{r}{The radial distances (km) at which the model should be evaluated.}

\item{a}{The azimuthal distances (rad) at which the model should be 
evaluated.}

\item{fit}{A model fit from \code{fit_sectorial}.}

\item{type}{Type of prediction: 'psf', 'dpsf', or 'cumpsf'. See Details.}

\item{tpred}{The level of the third predictor at which the model should be 
evaluated.}
}
\value{
A matrix with the PSF, density PSF or cumulative PSF.
}
\description{
Solves a fitted model from \code{fit_sectorial} for the PSF of the radial and 
azimuthal cummulative PSF at requested radius and azimuth, possibly including 
a third predictor variable.
}
\details{
If type = cumpsf, the model fit to the cumulative PSF will be 
evaluated at the desired positions. If type = dpsf, the density PSF is 
returned (1/km2). If type == psf, quadrature is used to calculate the average 
density PSF of the sector and the average is scaled by the area of the sector. 
This should equal the Monte Carlo results.  Note that in this case, r will be 
sorted and the returned values are for the mid points of the input vector of 
positions, so will have a length of length(r) - 1. Default is to return the 
PSF.
}
\examples{


}
\seealso{
\code{fit_sectorial}, \code{fit_annular}, \code{predict_annular}
}
